Extract Network and Interface from utils (#127)
* Extract Network and Interface from utils

* Use a component

* fmt

* Remove unnecessary dev-dep

* fmt

* extract setup_for_test


* Fmt


* Fixup failpoint test


* Fmt
