+++ harness/src/lib.rs
@@ -1,0 +1,49 @@
+// Copyright 2018 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Copyright 2015 The etcd Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+/*!
+
+This module contains various testing harness utilities for Raft.
+
+> If you want to build Raft without this, disable the `harness` feature.
+
+*/
+
+extern crate env_logger;
+extern crate raft;
+extern crate rand;
+
+mod interface;
+mod network;
+
+pub use self::{interface::Interface, network::Network};
+
+/// Do any common test initialization. Eg set up logging.
+#[doc(hidden)]
+pub fn setup_for_test() {
+    let _ = env_logger::try_init();
+}
