+++ tests/failpoint_cases/mod.rs
@@ -12,6 +12,7 @@
 // limitations under the License.
 
 use fail;
+use harness::setup_for_test;
 use raft::eraftpb::MessageType;
 use std::sync::*;
 use test_util::*;
