+++ tests/test_util/mod.rs
@@ -25,19 +25,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use env_logger;
+use harness::*;
 use protobuf::RepeatedField;
 use raft::eraftpb::*;
 use raft::storage::MemStorage;
 use raft::*;
-use rand;
-use std::collections::HashMap;
-use std::ops::*;
 
-/// Do any common test initialization. Eg set up logging, setup fail-rs.
-pub fn setup_for_test() {
-    let _ = env_logger::try_init();
-}
+#[allow(clippy::declare_interior_mutable_const)]
+pub const NOP_STEPPER: Option<Interface> = Some(Interface { raft: None });
 
 pub fn ltoa(raft_log: &RaftLog<MemStorage>) -> String {
     let mut s = format!("committed: {}\n", raft_log.committed);
@@ -66,76 +61,6 @@ pub fn new_test_config(
         max_size_per_msg: NO_LIMIT,
         max_inflight_msgs: 256,
         ..Default::default()
-    }
-}
-
-/// Compare to upstream, we use struct instead of trait here.
-/// Because to be able to cast Interface later, we have to make
-/// Raft derive Any, which will require a lot of dependencies to derive Any.
-/// That's not worthy for just testing purpose.
-pub struct Interface {
-    pub raft: Option<Raft<MemStorage>>,
-}
-
-impl From<Raft<MemStorage>> for Interface {
-    fn from(value: Raft<MemStorage>) -> Self {
-        Interface::new(value)
-    }
-}
-
-impl Interface {
-    pub fn new(r: Raft<MemStorage>) -> Interface {
-        Interface { raft: Some(r) }
-    }
-
-    pub fn step(&mut self, m: Message) -> Result<()> {
-        match self.raft {
-            Some(_) => Raft::step(self, m),
-            None => Ok(()),
-        }
-    }
-
-    pub fn read_messages(&mut self) -> Vec<Message> {
-        match self.raft {
-            Some(_) => self.msgs.drain(..).collect(),
-            None => vec![],
-        }
-    }
-
-    fn initial(&mut self, id: u64, ids: &[u64]) {
-        if self.raft.is_some() {
-            self.id = id;
-            let prs = self.take_prs();
-            self.set_prs(ProgressSet::with_capacity(
-                ids.len(),
-                prs.learner_ids().len(),
-            ));
-            for id in ids {
-                let progress = Progress::new(0, 256);
-                if prs.learner_ids().contains(id) {
-                    if let Err(e) = self.mut_prs().insert_learner(*id, progress) {
-                        panic!("{}", e);
-                    }
-                } else if let Err(e) = self.mut_prs().insert_voter(*id, progress) {
-                    panic!("{}", e);
-                }
-            }
-            let term = self.term;
-            self.reset(term);
-        }
-    }
-}
-
-impl Deref for Interface {
-    type Target = Raft<MemStorage>;
-    fn deref(&self) -> &Raft<MemStorage> {
-        self.raft.as_ref().unwrap()
-    }
-}
-
-impl DerefMut for Interface {
-    fn deref_mut(&mut self) -> &mut Raft<MemStorage> {
-        self.raft.as_mut().unwrap()
     }
 }
 
@@ -220,154 +145,4 @@ pub fn new_snapshot(index: u64, term: u64, nodes: Vec<u64>) -> Snapshot {
     s.mut_metadata().set_term(term);
     s.mut_metadata().mut_conf_state().set_nodes(nodes);
     s
-}
-
-#[derive(Default, Debug, PartialEq, Eq, Hash)]
-struct Connem {
-    from: u64,
-    to: u64,
-}
-
-#[allow(clippy::declare_interior_mutable_const)]
-pub const NOP_STEPPER: Option<Interface> = Some(Interface { raft: None });
-
-#[derive(Default)]
-pub struct Network {
-    pub peers: HashMap<u64, Interface>,
-    pub storage: HashMap<u64, MemStorage>,
-    dropm: HashMap<Connem, f64>,
-    ignorem: HashMap<MessageType, bool>,
-}
-
-impl Network {
-    // initializes a network from peers.
-    // A nil node will be replaced with a new *stateMachine.
-    // A *stateMachine will get its k, id.
-    // When using stateMachine, the address list is always [1, n].
-    pub fn new(peers: Vec<Option<Interface>>) -> Network {
-        Network::new_with_config(peers, false)
-    }
-
-    // new_with_config is like new but sets the configuration pre_vote explicitly
-    // for any state machines it creates.
-    pub fn new_with_config(mut peers: Vec<Option<Interface>>, pre_vote: bool) -> Network {
-        let size = peers.len();
-        let peer_addrs: Vec<u64> = (1..=size as u64).collect();
-        let mut nstorage = HashMap::new();
-        let mut npeers = HashMap::new();
-        for (p, id) in peers.drain(..).zip(peer_addrs.clone()) {
-            match p {
-                None => {
-                    nstorage.insert(id, new_storage());
-                    let r = new_test_raft_with_prevote(
-                        id,
-                        peer_addrs.clone(),
-                        10,
-                        1,
-                        nstorage[&id].clone(),
-                        pre_vote,
-                    );
-                    npeers.insert(id, r);
-                }
-                Some(mut p) => {
-                    p.initial(id, &peer_addrs);
-                    npeers.insert(id, p);
-                }
-            }
-        }
-        Network {
-            peers: npeers,
-            storage: nstorage,
-            ..Default::default()
-        }
-    }
-
-    pub fn ignore(&mut self, t: MessageType) {
-        self.ignorem.insert(t, true);
-    }
-
-    pub fn filter(&self, msgs: impl IntoIterator<Item = Message>) -> Vec<Message> {
-        msgs.into_iter()
-            .filter(|m| {
-                if self
-                    .ignorem
-                    .get(&m.get_msg_type())
-                    .cloned()
-                    .unwrap_or(false)
-                {
-                    return false;
-                }
-                // hups never go over the network, so don't drop them but panic
-                assert_ne!(m.get_msg_type(), MessageType::MsgHup, "unexpected msgHup");
-                let perc = self
-                    .dropm
-                    .get(&Connem {
-                        from: m.get_from(),
-                        to: m.get_to(),
-                    })
-                    .cloned()
-                    .unwrap_or(0f64);
-                rand::random::<f64>() >= perc
-            })
-            .collect()
-    }
-
-    pub fn read_messages(&mut self) -> Vec<Message> {
-        self.peers
-            .iter_mut()
-            .flat_map(|(_peer, progress)| progress.read_messages())
-            .collect()
-    }
-
-    pub fn send(&mut self, msgs: Vec<Message>) {
-        let mut msgs = msgs;
-        while !msgs.is_empty() {
-            let mut new_msgs = vec![];
-            for m in msgs.drain(..) {
-                let resp = {
-                    let p = self.peers.get_mut(&m.get_to()).unwrap();
-                    let _ = p.step(m);
-                    p.read_messages()
-                };
-                new_msgs.append(&mut self.filter(resp));
-            }
-            msgs.append(&mut new_msgs);
-        }
-    }
-
-    /// Dispatches the given messages to the appropriate peers.
-    ///
-    /// Unlike `send` this does not gather and send any responses. It also does not ignore errors.
-    pub fn dispatch(&mut self, messages: impl IntoIterator<Item = Message>) -> Result<()> {
-        for message in self.filter(messages) {
-            let to = message.get_to();
-            let peer = self.peers.get_mut(&to).unwrap();
-            peer.step(message)?;
-        }
-        Ok(())
-    }
-
-    pub fn drop(&mut self, from: u64, to: u64, perc: f64) {
-        self.dropm.insert(Connem { from, to }, perc);
-    }
-
-    pub fn cut(&mut self, one: u64, other: u64) {
-        self.drop(one, other, 1f64);
-        self.drop(other, one, 1f64);
-    }
-
-    pub fn isolate(&mut self, id: u64) {
-        for i in 0..self.peers.len() as u64 {
-            let nid = i + 1;
-            if nid != id {
-                self.drop(id, nid, 1.0);
-                self.drop(nid, id, 1.0);
-            }
-        }
-    }
-
-    pub fn recover(&mut self) {
-        self.dropm = HashMap::new();
-        self.ignorem = HashMap::new();
-    }
 }
